Move tuya models to separate module (#148550)
