+++ homeassistant/components/tuya/alarm_control_panel.py
@@ -20,7 +20,8 @@ from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
 
 from . import TuyaConfigEntry
 from .const import TUYA_DISCOVERY_NEW, DPCode, DPType
-from .entity import EnumTypeData, TuyaEntity
+from .entity import TuyaEntity
+from .models import EnumTypeData
 
 
 @dataclass(frozen=True)
